lambda/requirements.txt
aws_lambda_powertools==3.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
jmespath==1.0.1
joblib==1.4.2
nltk==3.9.1
praw==7.8.1
prawcore==2.4.0
regex==2024.11.6
requests==2.32.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.13.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
aws-xray-sdk
aws_lambda_powertools
praw
python-dateutil
requests
boto3
transformers
torch